btrfs: fix qgroup_free_reserved_data int overflow

+++ fs/btrfs/qgroup.c
@@ -4069,7 +4069,7 @@ static int qgroup_free_reserved_data(struct btrfs_inode *inode,
 	struct ulist_node *unode;
 	struct ulist_iterator uiter;
 	struct extent_changeset changeset;
-	int freed = 0;
+	u64 freed = 0;
 	int ret;
 
 	extent_changeset_init(&changeset);
